Fix compiler warnings about narrowing, name hiding, unused vairables or parameters. Silence some msvc warnings.
(cherry picked from commit 613108454d71288e20ed5c03d6b53400ba804284)

+++ include/boost/geometry/algorithms/detail/overlay/sort_by_side.hpp
@@ -299,7 +299,7 @@ struct side_sorter
             return;
         }
 
-        int const last = 1 + m_ranked_points.back().main_rank;
+        size_t const last = 1 + m_ranked_points.back().main_rank;
 
         // Move iterator after main_rank==0
         bool has_first = false;
